[#8] refactor: checkValidate 함수를 SignUpForm 컴포넌트 외부에 위치
