+++ src/utils/validate.ts
@@ -53,37 +53,3 @@ export const validatePasswordMatch = (
     return { isMathced: true, message: '비밀번호가 일치해요.' }
   }
 }
-
-export const checkValidate = (formValues: FormValues) => {
-  let errors: { [key: string]: { isMatched?: boolean; message: string } } = {}
-
-  if (formValues.email !== null) {
-    const result = validateEmail(formValues.email)
-    if (result) {
-      errors.email = result
-    }
-  }
-  if (formValues.password !== null) {
-    const result = validatePassword(formValues.password)
-    if (result) {
-      errors.password = result
-    }
-  }
-  if (formValues.confirmPassword !== null) {
-    const result = validatePasswordMatch(
-      formValues.password,
-      formValues.confirmPassword,
-    )
-    if (result) {
-      errors.confirmPassword = result
-    }
-  }
-  if (formValues.nickname !== null) {
-    const result = validateNickname(formValues.nickname)
-    if (result) {
-      errors.nickname = result
-    }
-  }
-
-  return errors
-}
